FLOW-965 FInputBase created for commn props and functions
