+++ packages/flow-core/src/components/f-input/f-input-global.scss
@@ -11,6 +11,7 @@ f-input {
 	width: 100%;
 	&[disabled] {
 		@include disabled();
+		opacity: 1;
 	}
 	@each $state, $color in $states {
 		&[state="#{$state}"] {
