+++ packages/flow-core/src/components/f-input/f-input-light-global.scss
