+++ packages/flow-core/src/index.ts
@@ -46,7 +46,7 @@ export * from "./components/f-breadcrumb/f-breadcrumb";
 export * from "./components/f-document-viewer/f-document-viewer";
 export * from "./mixins/components/f-root/f-root";
 export * from "./components/f-form-field/f-form-field";
-export * from "./components/f-input-light/f-input-light";
+export * from "./components/f-input/f-input-light";
 
 export { html } from "lit";
